add the token in password_policy and reason_code table

+++ src/controllers/reason-code-action.controller.ts
@@ -6,6 +6,7 @@ import { Module } from '../models/module.model';
 import { Op, where } from 'sequelize';
 import Event from '../models/event.model';
 import ReasonCodeModel from '../models/reason-code.model';
+import { decodeToken } from '../middlewares/verifyToken';
 
 
 export async function createReasoncode(
@@ -13,6 +14,16 @@ export async function createReasoncode(
     reply: FastifyReply
 ) {
     const traceId = generateCustomUUID();
+    const authHeader = request.headers.authorization;
+    if (!authHeader?.startsWith('Bearer')) {
+        return reply.status(401).send({ status_code: 401, message: 'Unauthorized-Token not found' });
+    }
+    const token = authHeader.split(' ')[1];
+    let user: any = await decodeToken(token);
+    if (!user) {
+        return reply.status(401).send({ status_code: 401, message: 'Unauthorized - Invalid token' });
+    }
+    const userId = user?.sub
     try {
         const reasoncode = request.body as {
             id: any;
@@ -48,7 +59,9 @@ export async function createReasoncode(
                 name: reason.name,
                 category: reason.category,
                 reason_code_id: reason_code_action.id,
-                is_enabled: reason.is_enabled
+                is_enabled: reason.is_enabled,
+                created_by: userId,
+                modified_by: userId,
             }))
         );
 
@@ -145,7 +158,7 @@ export async function getAllReasoncode(request: FastifyRequest, reply: FastifyRe
             reason_code_action: reasoncodesWithDetails,
             trace_id: traceId,
         });
-    } catch (error:any) {
+    } catch (error: any) {
         reply.status(500).send({
             status_code: 500,
             message: 'Internal server error',
@@ -332,6 +345,16 @@ export async function updateReasoncode(request: FastifyRequest, reply: FastifyRe
     const traceId = generateCustomUUID();
 
     const transaction = await ReasonCodeModel.sequelize?.transaction();
+    const authHeader = request.headers.authorization;
+    if (!authHeader?.startsWith('Bearer')) {
+        return reply.status(401).send({ status_code: 401, message: 'Unauthorized-Token not found' });
+    }
+    const token = authHeader.split(' ')[1];
+    let user: any = await decodeToken(token);
+    if (!user) {
+        return reply.status(401).send({ status_code: 401, message: 'Unauthorized - Invalid token' });
+    }
+    const userId = user?.sub
 
     try {
         await ReasonCodeModel.destroy({
@@ -354,6 +377,7 @@ export async function updateReasoncode(request: FastifyRequest, reply: FastifyRe
                     ...reasonCodeData,
                     reason_code_id: id,
                     program_id,
+                    modified_by: userId,
                 },
                 { transaction }
             );
